+++ src/schemas/versionedQuestionCondition.ts
@@ -1,0 +1,34 @@
+import gql from 'graphql-tag';
+
+export const typeDefs = gql`
+  extend type Query {
+    "Search for VersionedQuestions that belong to Section specified by sectionId"
+    publishedConditionsForQuestion(questionId: Int!): [VersionedQuestionCondition]
+  }
+
+  type VersionedQuestionCondition {
+    "The unique identifer for the Object"
+    id: Int
+    "The user who created the Object"
+    createdById: Int
+    "The timestamp when the Object was created"
+    created: DateTimeISO
+    "The user who last modified the Object"
+    modifiedById: Int
+    "The timestamp when the Object was last modifed"
+    modified: DateTimeISO
+    "Errors associated with the Object"
+    errors: [String!]
+
+    "The versionedQuestion id that the QuestionCondition belongs to"
+    versionedQuestionId: Int!
+    "The action to take on a QuestionCondition"
+    action: QuestionConditionActionType!
+    "The condition in which to take the action"
+    condition: QuestionConditionCondition!
+    "Relative to the condition type, it is the value to match on (e.g., HAS_ANSWER should equate to null here)"
+    conditionMatch: String
+    "The target of the action (e.g., an email address for SEND_EMAIL and a Question id otherwise)"
+    target: String!
+  }
+`;